# docs/swagger.yaml
basePath: /
definitions:
  controller.ActiveStreamResponse:
    properties:
      examid:
        example: EXAM001
        type: string
      interval:
        example: 5
        type: integer
      sessionid:
        example: abcdef1234567890
        type: string
      starttime:
        example: "2023-01-01T12:00:00Z"
        type: string
      streamtype:
        example: interval
        type: string
      userid:
        example: "2021001"
        type: string
      username:
        example: 张三
        type: string
    type: object
...
  controller.ErrorResponse:
    properties:
      error:
        example: 请求参数错误
        type: string
    type: object
...
  controller.InitScreenStreamRequest:
    properties:
      interval:
        description: 截屏间隔秒数（仅用于interval类型）
        example: 5
        type: integer
      streamtype:
        description: 流类型：interval-定时截屏, realtime-实时直播
        enum:
        - interval
        - realtime
        example: interval
        type: string
    type: object
  controller.InitScreenStreamResponse:
    properties:
      examid:
        example: EXAM001
        type: string
      interval:
        example: 5
        type: integer
      message:
        example: 屏幕流会话初始化成功
        type: string
      sessionid:
        example: abcdef1234567890
        type: string
      streamtype:
        example: interval
        type: string
      userid:
        example: "2021001"
        type: string
    type: object
  controller.RegisterUserRequest:
    properties:
      name:
        description: 用户姓名
        example: 张三
        type: string
      userid:
        description: 用户ID（学号/工号）
        example: "2021001"
        type: string
    required:
    - name
    - userid
    type: object
  controller.RegisterUserResponse:
    properties:
      message:
        example: 注册成功
        type: string
      name:
        example: 张三
        type: string
      userid:
        example: "2021001"
        type: string
    type: object
  controller.StopScreenStreamRequest:
    properties:
      sessionid:
        example: abcdef1234567890
        type: string
    required:
    - sessionid
    type: object
...
  controller.SuccessResponse:
    properties:
      message:
        example: 操作成功
        type: string
    type: object
...
  controller.User:
    properties:
      name:
        example: 张三
        type: string
      userid:
        example: "2021001"
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 提供考试密码生成、分发和验证功能的API系统
  title: 考试密码分发系统API
  version: "1.0"
paths:
...
  /screen/active:
    get:
      consumes:
      - application/json
      description: 管理员获取指定考试的所有活跃屏幕流
      parameters:
      - description: 考试ID
        in: query
        name: examid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.ActiveStreamResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取活跃的屏幕流
  /screen/init:
    post:
      consumes:
      - application/json
      description: 学生端初始化屏幕流会话
      parameters:
      - description: 初始化请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.InitScreenStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.InitScreenStreamResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 初始化屏幕流
  /screen/stop:
    post:
      consumes:
      - application/json
      description: 停止指定的屏幕流会话
      parameters:
      - description: 停止请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.StopScreenStreamRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 停止屏幕流
  /user/{userid}:
    get:
      consumes:
      - application/json
      description: 获取指定ID的用户信息
      parameters:
      - description: 用户ID
        in: path
        name: userid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.User'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 获取用户信息
  /user/register:
    post:
      consumes:
      - application/json
      description: 注册新用户（不关联考试，仅创建基本身份）
      parameters:
      - description: 注册信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.RegisterUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ErrorResponse'
      summary: 注册新用户
...
swagger: "2.0"